Custos sharing registry changes

+++ custos-sharing-registry-service/sharing-service-core/src/main/java/org/apache/custos/sharing/service/core/models/Sharing.java
@@ -8,8 +8,8 @@ public class Sharing {
     private SharingType sharingType;
     private String domainId;
     private String inheritedParentId;
-    private double createdTime;
-    private double updatedTime;
+    private Long createdTime;
+    private Long updatedTime;
 
     public String getPermissionTypeId() {
         return permissionTypeId;
@@ -59,19 +59,19 @@ public class Sharing {
         this.inheritedParentId = inheritedParentId;
     }
 
-    public double getCreatedTime() {
+    public Long getCreatedTime() {
         return createdTime;
     }
 
-    public void setCreatedTime(double createdTime) {
+    public void setCreatedTime(Long createdTime) {
         this.createdTime = createdTime;
     }
 
-    public double getUpdatedTime() {
+    public Long getUpdatedTime() {
         return updatedTime;
     }
 
-    public void setUpdatedTime(double updatedTime) {
+    public void setUpdatedTime(Long updatedTime) {
         this.updatedTime = updatedTime;
     }
